Pabaigta su i2c ir bme280

+++ inc/task_bme280.h
@@ -2,7 +2,7 @@
 #define TASK_BME280_H
 
 struct Task_BME280_Data {
-   short h, p, t;
+   int h, p, t;
 };
 
 void Task_BME280(void);
